requirements.txt
Flask==1.0.2
python-telegram-bot==11.1.0
requests==2.21.0
google-cloud-vision==0.35.2